Moved to `CustomerInfo`
