+++ Sources/Purchasing/EntitlementInfo.swift
@@ -224,7 +224,7 @@ extension PeriodType: DefaultValueProvider {
     ) {
         self.contents = .init(
             identifier: identifier,
-            isActive: Self.isDateActive(expirationDate: entitlement.expiresDate, for: requestDate),
+            isActive: CustomerInfo.isDateActive(expirationDate: entitlement.expiresDate, for: requestDate),
             willRenew: Self.willRenewWithExpirationDate(expirationDate: subscription.expiresDate,
                                                         store: subscription.store,
                                                         unsubscribeDetectedAt: subscription.unsubscribeDetectedAt,
@@ -284,37 +284,6 @@ public extension EntitlementInfo {
 
 // MARK: - Internal
 
-extension EntitlementInfo {
-
-    /// This grace period allows apps to continue functioning if the backend is down, but for a limited time.
-    /// We don't want to continue granting entitlements with an outdated `requestDate` forever,
-    /// since that would allow a user to get a free trial, then go offline and keep the entitlement with no time limit.
-    static let requestDateGracePeriod: DispatchTimeInterval = .days(3)
-
-}
-
-// MARK: - Private
-
-extension EntitlementInfo {
-
-    static func isDateActive(expirationDate: Date?, for requestDate: Date) -> Bool {
-        guard let expirationDate = expirationDate else {
-            return true
-        }
-
-        let (referenceDate, inGracePeriod) = Self.referenceDate(for: requestDate)
-        let isActive = expirationDate.timeIntervalSince(referenceDate) >= 0
-
-        if !inGracePeriod && !isActive {
-            Logger.warn(Strings.purchase.entitlement_expired_outside_grace_period(expiration: expirationDate,
-                                                                                  reference: requestDate))
-        }
-
-        return isActive
-    }
-
-}
-
 private extension EntitlementInfo {
 
     static func willRenewWithExpirationDate(expirationDate: Date?,
@@ -327,14 +296,6 @@ private extension EntitlementInfo {
         let hasBillingIssues = billingIssueDetectedAt != nil
 
         return !(isPromo || isLifetime || hasUnsubscribed || hasBillingIssues)
-    }
-
-    private static func referenceDate(for requestDate: Date) -> (Date, inGracePeriod: Bool) {
-        if Date().timeIntervalSince(requestDate) <= Self.requestDateGracePeriod.seconds {
-            return (requestDate, true)
-        } else {
-            return (Date(), false)
-        }
     }
 
 }
